Don't modify PATH if relevant tools are already present
The new 4.4 SDK from Core Devices includes a precompiled qemu-pebble binary and ARM toolchain. While this is useful for
most users, the binaries aren't usable on a NixOS system, due to the way NixOS handles dynamically-linked executables.
To work around this, pebble.nix provides binaries built/patched specifically for NixOS, for pebble-tool to use. However,
because the SDK binary paths are being appended to PATH, they get used first, overriding the versions pebble.nix
supplies, breaking app building and the emulator when using pebble.nix on NixOS.

To resolve this, a check is added to see if the relevant tools already exist in PATH before adding the SDK-provided
binaries. This way, any users that want to supply their own copies of qemu-pebble or the ARM toolchain for whatever
reason can do so easily, without pebble-tool forcing them to use binaries provided by the SDK.