Add environment variable to provide PATH overrides
The new 4.4 SDK from Core Devices includes a precompiled qemu-pebble binary and ARM toolchain. While this is useful for
most users, the binaries aren't usable on a NixOS system, due to the way NixOS handles dynamically-linked executables.
To work around this, pebble.nix provides binaries built/patched specifically for NixOS, for pebble-tool to use. However,
because the SDK binary paths are being appended to PATH, they get used first, overriding the versions pebble.nix
supplies, breaking app building and the emulator when using pebble.nix on NixOS.

To resolve this, we check for the existence of a PEBBLE_EXTRA_PATH environment variable, and append it to PATH if
present, overriding the SDK-provided entries. This way, any users that want to supply their own copies of qemu-pebble or
the ARM toolchain for whatever reason can do so easily, without pebble-tool forcing them to use binaries provided by the
SDK.